+++ source/DotNetActors.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HelloWorld", "HelloWorld\HelloWorld.csproj", "{BE05F2FC-E6D7-4DF1-A5E9-D3D2E821EA5A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestDotnetActorsWorking", "TestDotnetActorsWorking\TestDotnetActorsWorking.csproj", "{BE05F2FC-E6D7-4DF1-A5E9-D3D2E821EA5A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetActors", "DotNetActors\DotNetActors.csproj", "{813105BC-5CAE-4AEB-A5A2-8A1FAD802FAA}"
 EndProject
