+++ source/TestDotnetActorsWorking/TestDotnetActorsWorking.csproj
